@@ -12,18 +12,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.0.9",
-    "@angular/cdk": "6.4.0",
-    "@angular/common": "6.0.9",
-    "@angular/compiler": "6.0.9",
-    "@angular/core": "6.0.9",
-    "@angular/forms": "6.0.9",
-    "@angular/http": "6.0.9",
-    "@angular/material": "6.4.0",
-    "@angular/platform-browser": "6.0.9",
-    "@angular/platform-browser-dynamic": "6.0.9",
-    "@angular/router": "6.0.9",
-    "@angular/upgrade": "6.0.9",
+    "@angular/animations": "6.1.1",
+    "@angular/cdk": "6.4.3",
+    "@angular/common": "6.1.1",
+    "@angular/compiler": "6.1.1",
+    "@angular/core": "6.1.1",
+    "@angular/forms": "6.1.1",
+    "@angular/http": "6.1.1",
+    "@angular/material": "6.4.3",
+    "@angular/platform-browser": "6.1.1",
+    "@angular/platform-browser-dynamic": "6.1.1",
+    "@angular/router": "6.1.1",
+    "@angular/upgrade": "6.1.1",
     "classlist.js": "^1.1.20150312",
     "core-js": "2.5.7",
     "jasmine": "^3.1.0",
@@ -32,13 +32,13 @@
     "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "6.0.8",
-    "@angular/compiler-cli": "6.0.9",
-    "@angular/language-service": "6.0.9",
+    "@angular-devkit/build-angular": "0.7.1",
+    "@angular/cli": "6.1.2",
+    "@angular/compiler-cli": "6.1.1",
+    "@angular/language-service": "6.0.3",
     "@types/jasmine": "^2.8.8",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.4.2",
+    "codelyzer": "^4.4.3",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "^4.2.1",
     "karma": "~2.0.0",
@@ -49,6 +49,6 @@
     "protractor": "^5.4.0",
     "ts-node": "^7.0.0",
     "tslint": "^5.11.0",
-    "typescript": "2.7.2"
+    "typescript": "2.9.x"
   }
 }
