@@ -1,11 +1,11 @@
+import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpClientModule } from '@angular/common/http';
 
-import { PageComponent } from './page';
-import { EmployeeListComponent } from './employeeList';
 import { EmployeeDetailComponent } from './employeeDetail';
 import { EmployeeDisplayComponent } from './employeeDisplay';
+import { EmployeeListComponent } from './employeeList';
+import { PageComponent } from './page';
 
 @NgModule({
   declarations: [
